More Pattern Design
Had another test, and wanted all in the same place, so it's more accesible.

Gonna add memento, state and command.
Need to refactor to english so everybody understands the code, as well the description. Might add it in the readme, instead of writing another pdf, so it's easier to find the repo.

Coding is :art:

Shoutout to @brends31, he and me made all this repo possible.